+++ src/main/java/insane96mcp/enhancedai/mixin/WitchMixin.java
@@ -1,6 +1,7 @@
 package insane96mcp.enhancedai.mixin;
 
 import insane96mcp.enhancedai.modules.Modules;
+import insane96mcp.enhancedai.modules.witch.data.PotionOrMobEffect;
 import insane96mcp.enhancedai.modules.witch.feature.ThirstyWitches;
 import insane96mcp.enhancedai.modules.witch.feature.WitchPotionThrowing;
 import insane96mcp.enhancedai.setup.EAStrings;
@@ -8,8 +9,10 @@ import insane96mcp.insanelib.util.MCUtils;
 import net.minecraft.sounds.SoundEvents;
 import net.minecraft.tags.DamageTypeTags;
 import net.minecraft.tags.FluidTags;
+import net.minecraft.world.effect.MobEffect;
 import net.minecraft.world.effect.MobEffectInstance;
 import net.minecraft.world.effect.MobEffects;
+import net.minecraft.world.entity.EntityEvent;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.entity.ai.attributes.AttributeInstance;
@@ -27,6 +30,7 @@ import net.minecraft.world.item.Items;
 import net.minecraft.world.item.alchemy.PotionUtils;
 import net.minecraft.world.item.alchemy.Potions;
 import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.LevelEvent;
 import net.minecraft.world.phys.Vec3;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
@@ -35,8 +39,6 @@ import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.ModifyArg;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
-import java.util.ArrayList;
-import java.util.Collection;
 import java.util.List;
 import java.util.UUID;
 
@@ -86,8 +88,6 @@ public abstract class WitchMixin extends Raider {
 		this.healRaidersGoal.decrementCooldown();
 		this.attackPlayersGoal.setCanAttack(this.healRaidersGoal.getCooldown() <= 0);
 
-		Collection<MobEffectInstance> mobEffectInstances = new ArrayList<>();
-
 		if (this.isDrinkingPotion()) {
 			if (this.usingTime % 8 == 0)
 				this.playSound(SoundEvents.GENERIC_DRINK, 1.0f, this.random.nextFloat() * 0.1F + 0.9F);
@@ -109,41 +109,39 @@ public abstract class WitchMixin extends Raider {
 			}
 		}
 		else {
-			if (this.getTarget() != null && this.getTarget() instanceof Player && this.distanceToSqr(this.getTarget()) > 64d) {
-				List<MobEffectInstance> listToLoop = ThirstyWitches.drinkPotion;
-				for (MobEffectInstance mobEffectInstance : listToLoop) {
-					if (this.hasEffect(mobEffectInstance.getEffect()))
+			ItemStack stack = ItemStack.EMPTY;
+
+			if (this.getTarget() != null && this.getTarget() instanceof Player && this.distanceToSqr(this.getTarget()) > 36d) {
+				List<PotionOrMobEffect> listToLoop = ThirstyWitches.drinkPotion;
+				for (PotionOrMobEffect potionOrMobEffect : listToLoop) {
+					MobEffect mobEffect = potionOrMobEffect.getMobEffect();
+					if (mobEffect != null && this.hasEffect(mobEffect))
 						continue;
 
-					mobEffectInstances.add(new MobEffectInstance(mobEffectInstance));
+					stack = potionOrMobEffect.getPotionStack();
 					break;
 				}
 			}
 			else {
 				if (this.random.nextFloat() < ThirstyWitches.waterBreathingChance && this.isEyeInFluid(FluidTags.WATER) && !this.hasEffect(MobEffects.WATER_BREATHING) && this.getAirSupply() < this.getMaxAirSupply() / 2) {
-					mobEffectInstances.addAll(Potions.WATER_BREATHING.getEffects());
+					stack = PotionUtils.setPotion(new ItemStack(Items.POTION), Potions.WATER_BREATHING);
 				}
 				else if (this.random.nextFloat() < ThirstyWitches.fireResistanceChance && (this.isOnFire() || this.getLastDamageSource() != null && this.getLastDamageSource().is(DamageTypeTags.IS_FIRE)) && !this.hasEffect(MobEffects.FIRE_RESISTANCE)) {
-					mobEffectInstances.addAll(Potions.FIRE_RESISTANCE.getEffects());
+					stack = PotionUtils.setPotion(new ItemStack(Items.POTION), Potions.FIRE_RESISTANCE);
 				}
 				else if (this.random.nextFloat() < ThirstyWitches.healingChance && this.getHealth() < this.getMaxHealth() - 3) {
 					if (this.getHealth() < this.getMaxHealth() * ThirstyWitches.strongHealingThreshold)
-						mobEffectInstances.addAll(Potions.STRONG_HEALING.getEffects());
+						stack = PotionUtils.setPotion(new ItemStack(Items.POTION), Potions.STRONG_HEALING);
 					else
-						mobEffectInstances.addAll(Potions.HEALING.getEffects());
+						stack = PotionUtils.setPotion(new ItemStack(Items.POTION), Potions.HEALING);
 				}
 			}
 
-			ItemStack item = null;
-			if (!mobEffectInstances.isEmpty()) {
-				item = MCUtils.setCustomEffects(new ItemStack(Items.POTION), mobEffectInstances);
-			}
-			else if (MCUtils.hasLongNegativeEffect(this) && ThirstyWitches.shouldDrinkMilk(this.random)) {
-				item = new ItemStack(Items.MILK_BUCKET);
-			}
+			if (stack == null && MCUtils.hasLongNegativeEffect(this) && this.random.nextDouble() < ThirstyWitches.milkChance)
+				stack = new ItemStack(Items.MILK_BUCKET);
 
-			if (item != null) {
-				this.setItemSlot(EquipmentSlot.MAINHAND, item);
+			if (stack != ItemStack.EMPTY) {
+				this.setItemSlot(EquipmentSlot.MAINHAND, stack);
 				this.usingTime = this.getMainHandItem().getUseDuration();
 				this.setUsingItem(true);
 				if (!this.isSilent()) {
@@ -154,41 +152,42 @@ public abstract class WitchMixin extends Raider {
 				attributeinstance.removeModifier(SPEED_MODIFIER_DRINKING);
 				attributeinstance.addTransientModifier(SPEED_MODIFIER_DRINKING);
 			}
-		}
+			else {
+				if (WitchPotionThrowing.shouldUseSlowFalling() && this.fallDistance > 8 && !this.hasEffect(MobEffects.SLOW_FALLING)) {
+					ItemStack slowFallingStack = MCUtils.setCustomEffects(new ItemStack(Items.SPLASH_POTION), List.of(new MobEffectInstance(MobEffects.SLOW_FALLING, 300, 0)));
+					this.getLookControl().setLookAt(this.getX(), this.getY(), this.getZ());
+					if (!this.isSilent()) {
+						this.playSound(SoundEvents.WITCH_THROW, 1.0F, 0.8F + this.getRandom().nextFloat() * 0.4F);
+					}
+					this.level().levelEvent(LevelEvent.PARTICLES_SPELL_POTION_SPLASH, this.blockPosition(), PotionUtils.getColor(slowFallingStack));
+					List<MobEffectInstance> mobEffects = PotionUtils.getMobEffects(slowFallingStack);
+					for (MobEffectInstance mobEffect : mobEffects) {
+						this.addEffect(new MobEffectInstance(mobEffect));
+					}
+				}
 
-		if (WitchPotionThrowing.shouldUseSlowFalling() && this.fallDistance > 8 && !this.hasEffect(MobEffects.SLOW_FALLING)) {
-			ItemStack stack = MCUtils.setCustomEffects(new ItemStack(Items.SPLASH_POTION), List.of(new MobEffectInstance(MobEffects.SLOW_FALLING, 300, 0)));
-			this.getLookControl().setLookAt(this.getX(), this.getY(), this.getZ());
-			if (!this.isSilent()) {
-				this.playSound(SoundEvents.WITCH_THROW, 1.0F, 0.8F + this.getRandom().nextFloat() * 0.4F);
-			}
-			this.level().levelEvent(2002, this.blockPosition(), PotionUtils.getColor(stack));
-			List<MobEffectInstance> mobEffects = PotionUtils.getMobEffects(stack);
-			for (MobEffectInstance mobEffect : mobEffects) {
-				this.addEffect(new MobEffectInstance(mobEffect));
-			}
-		}
+				if (!this.hasEffect(MobEffects.INVISIBILITY) && --this.invisibilityCooldown <= 0 && this.getHealth() < this.getMaxHealth() * WitchPotionThrowing.healthThresholdInvisibility) {
+					ThrownPotion thrownPotion = new ThrownPotion(this.level(), this);
+					thrownPotion.setItem(MCUtils.setCustomEffects(new ItemStack(Items.SPLASH_POTION), List.of(new MobEffectInstance(MobEffects.INVISIBILITY, 200))));
+					thrownPotion.shoot(0, -1d, 0, 0.1f, 2f);
+					this.level().addFreshEntity(thrownPotion);
 
-		if (!this.hasEffect(MobEffects.INVISIBILITY) && --this.invisibilityCooldown <= 0 && this.getHealth() < this.getMaxHealth() * WitchPotionThrowing.healthThresholdInvisibility) {
-			ThrownPotion thrownPotion = new ThrownPotion(this.level(), this);
-			thrownPotion.setItem(MCUtils.setCustomEffects(new ItemStack(Items.SPLASH_POTION), List.of(new MobEffectInstance(MobEffects.INVISIBILITY, 200))));
-			thrownPotion.shoot(0, -1d, 0, 0.1f, 2f);
-			this.level().addFreshEntity(thrownPotion);
+					//Try 5 times to find a random spot
+					for (int i = 0; i < 5; i++) {
+						Vec3 randomPos = DefaultRandomPos.getPos(this, 16, 9);
+						if (randomPos != null) {
+							this.getNavigation().moveTo(randomPos.x, randomPos.y, randomPos.z, 1.1f);
+							break;
+						}
+					}
 
-			//Try 5 times to find a random spot
-			for (int i = 0; i < 5; i++) {
-				Vec3 randomPos = DefaultRandomPos.getPos(this, 16, 9);
-				if (randomPos != null) {
-					this.getNavigation().moveTo(randomPos.x, randomPos.y, randomPos.z, 1.1f);
-					break;
+					this.invisibilityCooldown = 20;
 				}
 			}
-
-			this.invisibilityCooldown = 20;
 		}
 
 		if (this.random.nextFloat() < 7.5E-4F) {
-			this.level().broadcastEntityEvent(this, (byte)15);
+			this.level().broadcastEntityEvent(this, EntityEvent.WITCH_HAT_MAGIC);
 		}
 
 		super.aiStep();